Add new docker image with minimal TVM configuration

+++ tests/scripts/task_cpp_unittest.sh
@@ -45,20 +45,23 @@ python3 tests/scripts/task_build.py \
     --cmake-target cpptest \
     --build-dir "${BUILD_DIR}"
 
-# crttest requires USE_MICRO to be enabled, which is currently the case
-# with all CI configs
-pushd "${BUILD_DIR}"
-ninja crttest
-popd
+# crttest requires USE_MICRO to be enabled.
+if grep -Fq "set(USE_MICRO ON)" ${BUILD_DIR}/config.cmake; then
+  pushd "${BUILD_DIR}"
+  ninja crttest
+  popd
+fi
 
 
 pushd "${BUILD_DIR}"
 ctest --gtest_death_test_style=threadsafe
 popd
 
-# Test MISRA-C runtime
-pushd apps/bundle_deploy
-rm -rf build
-make test_dynamic test_static
-popd
+# Test MISRA-C runtime. It requires USE_MICRO to be enabled.
+if grep -Fq "set(USE_MICRO ON)" ${BUILD_DIR}/config.cmake; then
+  pushd apps/bundle_deploy
+  rm -rf build
+  make test_dynamic test_static
+  popd
+fi
